Created init_window
